clean(test): move the runner test abstraction to core to use it for jdbc and kafka
